backend/foodgram/requirements.txt
requests==2.26.0
django==2.2.16
django-filter==21.1
djangorestframework==3.11.2
djangorestframework-simplejwt==5.2.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
drf_yasg==1.21.3
django_filter
python-dotenv
gunicorn==20.0.4
psycopg2-binary==2.8.6
drf-base64==2.0
djoser==2.0.1
reportlab==3.6.3
python-dotenv==0.20.0
fpdf==1.7.2
sqlparse==0.4.2

Pillow==9.0.1